feat: add reader to encoder

+++ firmware/include/types/telemetry_types.h
@@ -6,8 +6,8 @@
 struct DeviceStatus
 {
     DeviceState state;
-    uint32_t angle = 0;
-    uint16_t center = 0;
+    int16_t angle = 0;
+    int16_t center = 0;
     uint8_t output = 0;
     uint8_t limit = 0;
     uint8_t throttle = 0;
